+++ tests/unit/test_athena.py
@@ -25,7 +25,6 @@ from .._utils import (
     get_df_txt,
     get_time_str_with_random_suffix,
     pandas_equals,
-    ts,
 )
 
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
@@ -1487,361 +1486,3 @@ def test_athena_date_recovery(path, glue_database, glue_table):
         ctas_approach=False,
     )
     assert pandas_equals(df, df2)
-
-
-@pytest.mark.parametrize("partition_cols", [None, ["name"], ["name", "day(ts)"]])
-@pytest.mark.parametrize(
-    "additional_table_properties",
-    [None, {"write_target_data_file_size_bytes": 536870912, "optimize_rewrite_delete_file_threshold": 10}],
-)
-def test_athena_to_iceberg(path, path2, glue_database, glue_table, partition_cols, additional_table_properties):
-    df = pd.DataFrame(
-        {
-            "id": [1, 2, 3],
-            "name": ["a", "b", "c"],
-            "ts": [ts("2020-01-01 00:00:00.0"), ts("2020-01-02 00:00:01.0"), ts("2020-01-03 00:00:00.0")],
-        }
-    )
-    df["id"] = df["id"].astype("Int64")  # Cast as nullable int64 type
-    df["name"] = df["name"].astype("string")
-
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        partition_cols=partition_cols,
-        additional_table_properties=additional_table_properties,
-    )
-
-    df_out = wr.athena.read_sql_query(
-        sql=f'SELECT * FROM "{glue_table}" ORDER BY id',
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-
-    assert df.equals(df_out)
-
-
-def test_athena_to_iceberg_schema_evolution_add_columns(
-    path: str, path2: str, glue_database: str, glue_table: str
-) -> None:
-    df = pd.DataFrame({"c0": [0, 1, 2], "c1": [3, 4, 5]})
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-    )
-
-    df["c2"] = [6, 7, 8]
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-    )
-
-    column_types = wr.catalog.get_table_types(glue_database, glue_table)
-    assert len(column_types) == len(df.columns)
-
-    df_out = wr.athena.read_sql_table(
-        table=glue_table,
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-    assert len(df_out) == len(df) * 2
-
-    df["c3"] = [9, 10, 11]
-    with pytest.raises(wr.exceptions.InvalidArgumentValue):
-        wr.athena.to_iceberg(
-            df=df,
-            database=glue_database,
-            table=glue_table,
-            table_location=path,
-            temp_path=path2,
-            keep_files=False,
-            schema_evolution=False,
-        )
-
-
-def test_athena_to_iceberg_schema_evolution_modify_columns(
-    path: str, path2: str, glue_database: str, glue_table: str
-) -> None:
-    # Version 1
-    df = pd.DataFrame({"c1": pd.Series([1.0, 2.0], dtype="float32"), "c2": pd.Series([-1, -2], dtype="int32")})
-
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-    )
-
-    df_out = wr.athena.read_sql_table(
-        table=glue_table,
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-
-    assert len(df_out) == 2
-    assert len(df_out.columns) == 2
-    assert str(df_out["c1"].dtype).startswith("float32")
-    assert str(df_out["c2"].dtype).startswith("Int32")
-
-    # Version 2
-    df2 = pd.DataFrame({"c1": pd.Series([3.0, 4.0], dtype="float64"), "c2": pd.Series([-3, -4], dtype="int64")})
-
-    wr.athena.to_iceberg(
-        df=df2,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-    )
-
-    df2_out = wr.athena.read_sql_table(
-        table=glue_table,
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-
-    assert len(df2_out) == 4
-    assert len(df2_out.columns) == 2
-    assert str(df2_out["c1"].dtype).startswith("float64")
-    assert str(df2_out["c2"].dtype).startswith("Int64")
-
-
-def test_athena_to_iceberg_schema_evolution_fill_missing_columns(
-    path: str, path2: str, glue_database: str, glue_table: str
-) -> None:
-    df = pd.DataFrame({"c0": [0, 1, 2], "c1": ["foo", "bar", "baz"]})
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-    )
-
-    print(wr.catalog.table(glue_database, glue_table))
-
-    df = pd.DataFrame({"c0": [3, 4, 5]})
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-        schema_fill_missing=True,
-    )
-    print(wr.catalog.table(glue_database, glue_table))
-
-    df_actual = wr.athena.read_sql_table(
-        table=glue_table,
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-    df_actual = df_actual.sort_values("c0").reset_index(drop=True)
-    df_actual["c0"] = df_actual["c0"].astype("int64")
-
-    df_expected = pd.DataFrame({"c0": [0, 1, 2, 3, 4, 5], "c1": ["foo", "bar", "baz", np.nan, np.nan, np.nan]})
-    df_expected["c1"] = df_expected["c1"].astype("string")
-
-    assert_pandas_equals(df_actual, df_expected)
-
-
-def test_athena_to_iceberg_schema_evolution_drop_columns_error(
-    path: str, path2: str, glue_database: str, glue_table: str
-) -> None:
-    df = pd.DataFrame({"c0": [0, 1, 2], "c1": [3, 4, 5]})
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        schema_evolution=True,
-    )
-
-    df = pd.DataFrame({"c0": [6, 7, 8]})
-
-    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
-        wr.athena.to_iceberg(
-            df=df,
-            database=glue_database,
-            table=glue_table,
-            table_location=path,
-            temp_path=path2,
-            keep_files=False,
-            schema_evolution=True,
-        )
-
-
-def test_to_iceberg_cast(path, path2, glue_table, glue_database):
-    df = pd.DataFrame(
-        {
-            "c0": [
-                datetime.date(4000, 1, 1),
-                datetime.datetime(2000, 1, 1, 10),
-                "2020",
-                "2020-01",
-                1,
-                None,
-                pd.NA,
-                pd.NaT,
-                np.nan,
-                np.inf,
-            ]
-        }
-    )
-    df_expected = pd.DataFrame(
-        {
-            "c0": [
-                datetime.date(1970, 1, 1),
-                datetime.date(2000, 1, 1),
-                datetime.date(2020, 1, 1),
-                datetime.date(2020, 1, 1),
-                datetime.date(4000, 1, 1),
-                None,
-                None,
-                None,
-                None,
-                None,
-            ]
-        }
-    )
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        dtype={"c0": "date"},
-    )
-    df2 = wr.athena.read_sql_table(database=glue_database, table=glue_table, ctas_approach=False)
-    assert pandas_equals(df_expected, df2.sort_values("c0").reset_index(drop=True))
-
-
-def test_athena_to_iceberg_with_hyphenated_table_name(
-    path: str, path2: str, glue_database: str, glue_table_with_hyphenated_name: str
-):
-    df = pd.DataFrame({"c0": [1, 2, 3, 4], "c1": ["foo", "bar", "baz", "boo"]})
-    df["c0"] = df["c0"].astype("int")
-    df["c1"] = df["c1"].astype("string")
-
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table_with_hyphenated_name,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-    )
-
-    df_out = wr.athena.read_sql_query(
-        sql=f'SELECT * FROM "{glue_table_with_hyphenated_name}"',
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-
-    assert len(df) == len(df_out)
-    assert len(df.columns) == len(df_out.columns)
-
-
-def test_athena_to_iceberg_column_comments(path: str, path2: str, glue_database: str, glue_table: str) -> None:
-    df = pd.DataFrame({"c0": [0, 1, 2], "c1": [3, 4, 5]})
-    column_comments = {
-        "c0": "comment 0",
-        "c1": "comment 1",
-    }
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        glue_table_settings={
-            "columns_comments": column_comments,
-        },
-    )
-
-    column_comments_actual = wr.catalog.get_columns_comments(glue_database, glue_table)
-
-    assert column_comments_actual == column_comments
-
-
-def test_athena_to_iceberg_merge_into(path: str, path2: str, glue_database: str, glue_table: str) -> None:
-    df = pd.DataFrame({"title": ["Dune", "Fargo"], "year": ["1984", "1996"], "gross": [35_000_000, 60_000_000]})
-    df["title"] = df["title"].astype("string")
-    df["year"] = df["year"].astype("string")
-    df["gross"] = df["gross"].astype("Int64")
-
-    wr.athena.to_iceberg(
-        df=df,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-    )
-
-    # Perform MERGE INTO
-    df2 = pd.DataFrame({"title": ["Dune", "Fargo"], "year": ["2021", "1996"], "gross": [400_000_000, 60_000_001]})
-    df2["title"] = df2["title"].astype("string")
-    df2["year"] = df2["year"].astype("string")
-    df2["gross"] = df2["gross"].astype("Int64")
-
-    wr.athena.to_iceberg(
-        df=df2,
-        database=glue_database,
-        table=glue_table,
-        table_location=path,
-        temp_path=path2,
-        keep_files=False,
-        merge_cols=["title", "year"],
-    )
-
-    # Expected output
-    df_expected = pd.DataFrame(
-        {
-            "title": ["Dune", "Fargo", "Dune"],
-            "year": ["1984", "1996", "2021"],
-            "gross": [35_000_000, 60_000_001, 400_000_000],
-        }
-    )
-    df_expected["title"] = df_expected["title"].astype("string")
-    df_expected["year"] = df_expected["year"].astype("string")
-    df_expected["gross"] = df_expected["gross"].astype("Int64")
-
-    df_out = wr.athena.read_sql_query(
-        sql=f'SELECT * FROM "{glue_table}" ORDER BY year',
-        database=glue_database,
-        ctas_approach=False,
-        unload_approach=False,
-    )
-
-    assert_pandas_equals(df_expected, df_out)
